% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARI.R
\docType{data}
\name{ARI}
\alias{ARI}
\title{ARI data from MEPARI trial}
\format{
See References.
}
\usage{
ARI
}
\description{
Acute Resipratory Infection (ARI) data from Meditation or Exercise for Preventing Acute Respiratory Infection (MEPARI) trial.
}
\references{
Full data package:

\url{https://www.openicpsr.org/openicpsr/project/103581/version/V2/view}

Direct link to the .csv file:

\url{https://www.openicpsr.org/openicpsr/project/103581/version/V2/view?path=/openicpsr/103581/fcr:versions/V2/ARI-Data/MEPARI1-2_ari_2017_1219.csv&type=file}
}
\seealso{
\url{https://shahin-roshani.github.io/BC2/}
}
\keyword{datasets}
