% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc2.R
\name{bc2}
\alias{bc2}
\title{bc2}
\usage{
bc2(
  pos.resps,
  preds,
  data,
  search.space = NULL,
  control.pars = NULL,
  parallel = F,
  core.nums = NULL,
  g.funs = NULL
)
}
\arguments{
\item{pos.resps}{A character vector of length two containing the names of model's two responses.}

\item{preds}{A character vector containing the names of predictors to be used in all parts of the model.}

\item{data}{Data containing responses and predictors.}

\item{search.space}{A named list containing usual parameters for controlling the optimization process of log-likelihood function.

All combinations of the given values in the list will be considered. Valid arguments inside the list are:

\itemize{
\item coeffs.start: Numeric vector containing values to be used as starting points of intercepts (a0, a1, a2), coefficients and constraints (b1 & b2) during optimization.
\item sigmas.start: Numeric vector containing values to be used as starting points of sigma1 & sigma2 during optimization.
\item ro.start: Numeric vector containing values to be used as starting points of ro during optimization.
\item methods: Optimization methods to be tested for optimizing the log-likelihood function.
\item iterations: Number of iterations to achieve the convergence in optimization process.
}

Default is NULL where a pre-defined search space will be used. See part 1 in Details.}

\item{control.pars}{Additional control parameters used by optim function. Default is NULL. See Details section in \code{?optim}.}

\item{parallel}{Logical indicating whether optimization of different combinations in search.space must be done in parallel mode or not. Default is NULL.}

\item{core.nums}{Number of CPU cores to be used for parallel computation. Default is NULL where half of the CPU cores will be used. See \code{?detectCores}.}

\item{g.funs}{A single function or a list of length two containing the functions to be used for transforming non-zero responses.

If \code{g.funs} were specified by user, $Data part of return object will contain the un-transformed responses while structured data part will include transformed responses based on functions in \code{g.funs}.}
}
\value{
S3 print class of method 'bc2' which is A list of length 3 containing fitting information of the logistic and the two positive parts.

This list is a part of a more comprehensive return that can be accessed when the fit is saved in an object. The full return set contains:

\itemize{
\item $Data: Cleaned data-set with original and un-transformed responses. Also see \code{g.funs} argument.
\item $`Structured data`: A list of divided and cleaned data to be used in the fitting process. If \code{g.funs} were specified, transformed responses will be included in this part.
\item $`Optimization info`: A tibble containing information about evaluation of different combinations that was given in search.space for optimization.
\item $`Best combination`: The best combination of parameters that results in the best set of estimations.
\item $`Final tables`: The default print output of class bc2 that was mentioned earlier.
}
}
\description{
Function for fitting Multivariate semi-continuous proportionally constrained two-part fixed effects model.
}
\details{
When search.space is NULL, a pre-defined search space will be used which is:

\preformatted{
list(coeffs.start=seq(-1,1,1),

     sigmas.start=c(.1,1),

     ro.start=c(-.5,.5),

     methods=c('Nelder-Mead','BFGS'),

     iterations=c(100,500))
}
}
\references{
\url{https://journals.sagepub.com/doi/abs/10.1177/0962280218807730}
}
\author{
Shahin Roshani
}
