% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMmvpolySelfDesign (1).R,
%   R/EMmvpolySelfDesign.R
\name{EMmvpolySelfDesign}
\alias{EMmvpolySelfDesign}
\alias{EMmvpolySelfDesign}
\title{Title}
\usage{
EMmvpolySelfDesign(y, x.self.design, z.design, baselineonly = NULL,
  additive = NULL, pairwise.interaction = NULL, saturated = NULL,
  missingTumorIndicator = 888, z.all = NULL)

EMmvpolySelfDesign(y, x.self.design, z.design, baselineonly = NULL,
  additive = NULL, pairwise.interaction = NULL, saturated = NULL,
  missingTumorIndicator = 888, z.all = NULL)
}
\arguments{
\item{z.all}{}

\item{z.all}{}
}
\description{
Title

Title
}
