% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMmvpolySelfDesignnew.R
\name{EMmvpolySelfDesignnew}
\alias{EMmvpolySelfDesignnew}
\title{Title}
\usage{
EMmvpolySelfDesignnew(y, x.self.design, z.design, x.self.design2, z.design2,
  baselineonly = NULL, additive = NULL, pairwise.interaction = NULL,
  saturated = NULL, missingTumorIndicator = 888, z.all = NULL,
  delta0 = NULL)
}
\arguments{
\item{delta0}{}
}
\description{
Title
}
