% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.bc2.R
\name{fitted.bc2}
\alias{fitted.bc2}
\title{fitted.bc2}
\usage{
\method{fitted}{bc2}(object, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "bc2".}

\item{...}{Other fitted arguments (Not being used by objects of class \code{bc2}).}
}
\value{
A tibble with 2 columns containing fitted values of each response.
}
\description{
fitted method for bc2 fits.
}
\seealso{
\url{https://shahin-roshani.github.io/BC2/}
}
\author{
Shahin Roshani
}
