% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCovarNamenew.R
\name{GenerateCovarNamenew}
\alias{GenerateCovarNamenew}
\title{Title}
\usage{
GenerateCovarNamenew(baselineonly, additive, pairwise.interaction, saturated,
  x.self.design, x.self.design2)
}
\arguments{
\item{x.self.design2}{}
}
\description{
Title
}
