% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSecondStageMat.R
\name{GenerateSecondStageMat}
\alias{GenerateSecondStageMat}
\title{Title}
\usage{
GenerateSecondStageMat(baselineonly.number, main.effect.number,
  pairwise.interaction.number, saturated.number, baselineonly.second.cat,
  main.effect.second.cat, pairwise.interaction.second.cat, saturated.second.cat,
  M, total.covar.number, full.second.stage.names, covar.names, delta)
}
\arguments{
\item{covar.names}{}
}
\description{
Title
}
