% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSecondStageTest.R
\name{GenerateSecondStageTest}
\alias{GenerateSecondStageTest}
\title{Title}
\usage{
GenerateSecondStageTest(delta, sigma, M, second.stage.mat)
}
\arguments{
\item{second.stage.mat}{}
}
\description{
Title
}
