% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSelfSecondStageMat.R,
%   R/GenerateSelfSecondStageMatnew.R
\name{GenerateSelfSecondStageMat}
\alias{GenerateSelfSecondStageMat}
\alias{GenerateSelfSecondStageMat}
\title{Title}
\usage{
GenerateSelfSecondStageMat(x.self.design, z.design, M, full.second.stage.names,
  delta)

GenerateSelfSecondStageMat(x.self.design, z.design, M, full.second.stage.names,
  delta)
}
\arguments{
\item{delta}{}

\item{delta}{}
}
\description{
Title

Title
}
