% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSelfXALL.R, R/GenerateSelfXALLNew.R
\name{GenerateSelfXAll}
\alias{GenerateSelfXAll}
\alias{GenerateSelfXAll}
\title{Title}
\usage{
GenerateSelfXAll(y, x.self.design, x.self.design2, baselineonly, additive,
  pairwise.interaction, saturated)

GenerateSelfXAll(y, x.self.design, x.self.design2, baselineonly, additive,
  pairwise.interaction, saturated)
}
\arguments{
\item{saturated}{}

\item{saturated}{}
}
\description{
Title

Title
}
