% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_z_design_matrix.R
\name{GenerateZDesignNamesMainEffect}
\alias{GenerateZDesignNamesMainEffect}
\title{Title}
\usage{
GenerateZDesignNamesMainEffect(tumor.names)
}
\arguments{
\item{tumor.names}{}
}
\description{
Title
}
