% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialSetupnew.R
\name{InitialSetupnew}
\alias{InitialSetupnew}
\title{Title}
\usage{
InitialSetupnew(y.pheno.complete, baselineonly, additive, pairwise.interaction,
  saturated, x.self.design, z.design, x.self.design2, z.design2)
}
\arguments{
\item{z.design2}{}
}
\description{
Title
}
