% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpoly.R, R/mvpoly_a.R
\name{Mvpoly}
\alias{Mvpoly}
\alias{Mvpoly}
\title{Title}
\usage{
Mvpoly(delta0, y, x, z_design)

Mvpoly(delta0, y, x, z_design)
}
\arguments{
\item{delta0}{this is the starting value for all the second stage parameters.
You need to assign a starting value to make the iteration algorithm begin}

\item{y}{disease status matrix y; This is an N by M matrix; N is the total number of samples; M is the total number of possible categories;}

\item{x}{matrix, which contains all the covariates you want to fit into the model;}

\item{z_design}{This matrix maps the relationship between the first stage parameters beta
and second stage parameters theta for a specific covariate.
 Notice:For a specific covariate! You don't need to write the matrix between
 first stage paramters beta and second stage paramters theta for all the covariate.
 The function will generate everything for you. The function use the same design matrix for every covariate in the model;For example, for the HER+-,ER+-,PR+- case, the z matrix is: the z_design matrix in simulation1.R
 #This is a shortcoming of this version of code. In the furture,
 I could develop the version that user could specific different z matrix for different covariates.}

\item{delta0}{this is the starting value for all the second stage parameters.
You need to assign a starting value to make the iteration algorithm begin}

\item{y}{disease status matrix y; This is an N by M matrix; N is the total number of samples; M is the total number of possible categories;}

\item{x}{matrix, which contains all the covariates you want to fit into the model;}

\item{z_design}{This matrix maps the relationship between the first stage parameters beta
and second stage parameters theta for a specific covariate.
 Notice:For a specific covariate! You don't need to write the matrix between
 first stage paramters beta and second stage paramters theta for all the covariate.
 The function will generate everything for you. The function use the same design matrix for every covariate in the model;For example, for the HER+-,ER+-,PR+- case, the z matrix is: the z_design matrix in simulation1.R
 #This is a shortcoming of this version of code. In the furture,
 I could develop the version that user could specific different z matrix for different covariates.}
}
\description{
Title

Title
}
