% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bc2.R
\name{plot.bc2}
\alias{plot.bc2}
\title{plot.bc2}
\usage{
\method{plot}{bc2}(object, type = "residuals")
}
\arguments{
\item{object}{a fitted object of class inheriting from "bc2".}

\item{type}{Type of plot to be drawn. Valid inputs are \code{'responses'}, \code{'residuals'} or \code{'fit'}. Default is \code{'residuals'}.}
}
\value{
When \code{type='responses'}, density plots of original and un-transformed responses (non-zero) will be returned (unless transformation was done directly on non-zero responses and not through the \code{g.funs} arguemt).

When \code{type='residuals'}, density plot of residuals plus residuals vs fitted values plot for each response will be returned.

For \code{type='fit'}, actual vs fitted values of each response will be returned.
}
\description{
plot method for bc2 fits.
}
\seealso{
\url{https://shahin-roshani.github.io/BC2/}
}
\author{
Shahin Roshani
}
