% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.bc2.R
\name{residuals.bc2}
\alias{residuals.bc2}
\title{residuals.bc2}
\usage{
\method{residuals}{bc2}(object)
}
\arguments{
\item{object}{a fitted object of class inheriting from "bc2".}
}
\value{
A tibble with 2 columns containing residuals related to each response.
}
\description{
residuals (resid) method for bc2 fits.
}
\seealso{
\url{https://shahin-roshani.github.io/BC2/}
}
\author{
Shahin Roshani
}
