% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTestSupportMixedModelSelfDesign.R,
%   R/ScoreTestSupportSelfDesign.R
\name{ScoreTestSupportMixedModelSelfDesign}
\alias{ScoreTestSupportMixedModelSelfDesign}
\title{Title}
\usage{
ScoreTestSupportMixedModelSelfDesign(
  y,
  x.self.design,
  z.design,
  baselineonly = NULL,
  additive = NULL,
  pairwise.interaction = NULL,
  saturated = NULL,
  missingTumorIndicator = 888,
  delta0 = NULL
)

ScoreTestSupportMixedModelSelfDesign(
  y,
  x.self.design,
  z.design,
  baselineonly = NULL,
  additive = NULL,
  pairwise.interaction = NULL,
  saturated = NULL,
  missingTumorIndicator = 888,
  delta0 = NULL
)
}
\arguments{
\item{delta0}{}
}
\description{
Title

Title
}
