% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStageModel.R
\name{TwoStageModel}
\alias{TwoStageModel}
\title{Title}
\usage{
TwoStageModel(y, baselineonly = NULL, additive = NULL,
  pairwise.interaction = NULL, saturated = NULL,
  missingTumorIndicator = NULL, missingDataAlgorithm = "EM")
}
\arguments{
\item{missingDataAlgorithm}{}
}
\description{
Title
}
