% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZSelfDesigntoZallnew.R
\name{ZSelfDesigntoZallnew}
\alias{ZSelfDesigntoZallnew}
\title{Title}
\usage{
ZSelfDesigntoZallnew(x.self.design, x.self.design2, baselineonly, additive,
  pairwise.interaction, saturated, z.design, z.design2, z.design.baselineonly,
  z.design.additive, z.design.pairwise.interaction, z.design.saturated)
}
\arguments{
\item{z.design.saturated}{}
}
\description{
Title
}
