% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSecondStageMat.R
\name{GenerateSecondStageMat}
\alias{GenerateSecondStageMat}
\title{Title}
\usage{
GenerateSecondStageMat(baselineonly, main.effect, pairwise.interaction,
  saturated, M, full.second.stage.names, covar.names, delta,
  z.design.main.effect, z.design.pairwise.interaction, z.design.saturated)
}
\arguments{
\item{covar.names}{}
}
\description{
Title
}
