% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSelfSecondStageMat.R
\name{GenerateSelfSecondStageMat}
\alias{GenerateSelfSecondStageMat}
\title{Title}
\usage{
GenerateSelfSecondStageMat(z.design, x, M, full.second.stage.names, covar.names,
  delta)
}
\arguments{
\item{delta}{}
}
\description{
Title
}
